Update for exercise 4.21

+++ controllers/blogs.js
@@ -36,8 +36,20 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+    return null;
+  } else {
+    return response.status(401).json({ error: "token invalid" });
+  }
 });
 
 blogsRouter.put("/:id", async (request, response) => {
